Add note about the default PubSub implementation
Based on [this StackOverflow answer](https://stackoverflow.com/a/43790449/2115623) by @helfer I've added a pretty important note to the README that the default `PubSub` implementation should not be used in production.

I did not realise this and have been running it in production across multiple instances for months now. We've had memory leak warnings and all that, but couldn't pinpoint the source until I randomly stumbled upon that StackOverflow answer. In hindsight it's obvious, but when you're in the middle of a huge codebase with complex real-time requirements it's not obvious at all.