feat: replace iterall with native Symbol.asyncIterator + fix return types (#232)
* fix(typings): return AsyncIterableIterator instead of AsyncIterator

BREAKING fixes the type annotation of the abstract class PubSubEngine. According to the TypeScript type-defintion a `PubSubAsyncIterator` instance is actually a `AsyncIterableIterator` instead of an  `AsyncIterator`. The typing of `PubSubAsyncIterator` is way more convenient as it allows iterating over it with the `for await (const foo of iterator) { doSth() }` syntax, which is super handy for filtering or mapping (See https://gist.github.com/n1ru4l/127178705cc0942cad0e45d425e2eb63 for some example operators).

* remove iterall

* rename asyncIterator method to asyncIterableIterator.

* Slight tweaks based on graphql-js 16 changes

* Changelog update

Co-authored-by: hwillson <hugh@octonary.com>